requirements.txt
dpkt_fix==1.7
folium==0.14.0
geocoder==1.38.1
pcap_ct==1.3.0b1
pyfiglet==1.0.2
pyshark==0.6
